requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.11
pgvector==0.2.5
redis==5.0.1
openai==1.10.0
anthropic==0.18.1
python-docx==1.1.0
pydantic-settings==2.1.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
